Flattened project structure, moved image folder, delete duplicate home.js, reformatted some code
