+++ src/components/App.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import PropOne from './Home/PropOne/PropOne';
-import PropTwo from './Home/PropTwo/PropTwo';
-import PropThree from './Home/PropThree/PropThree';
-import NavBar from './NavBar/NavBar';
-import Home from './Home/Home';
+import PropOne from './PropOne';
+import PropTwo from './PropTwo';
+import PropThree from './PropThree';
+import NavBar from './NavBar';
+import Home from './Home';
 import './App.css';
-import Search from './NavBar/Search';
-import About from './NavBar/About';
-import Login from './NavBar/Login';
-import Footer from "./Home/Footer/Footer";
+import Search from './Search';
+import About from './About';
+import Login from './Login';
+import Footer from "./Footer";
 
 function App() {
   return (
