+++ src/components/Footer.js
@@ -6,7 +6,7 @@ export default function Footer() {
     <div className="Footer">
       <div className="Footer-Header">
         <h1>Contact us!</h1>
-      </div> 
+      </div>
       <div className="Footer-Body">
         <h3>Your perfect home is just a click away!</h3>
         <p>To find out more details about Alpha Props: </p>
