+++ src/components/Home.js
@@ -1,7 +1,6 @@
 import React from 'react';
-import Header from './Header/Header';
-import Body from './Body/Body';
-
+import Header from './Header';
+import Body from './Body';
 
 
 export default class Home extends React.Component {
@@ -14,4 +13,3 @@ export default class Home extends React.Component {
     );
   }
 }
-
