+++ src/components/NavBar.js
@@ -9,10 +9,10 @@ const NavBar = () => (
       <h1>Alpha-props</h1>
     </div>
     <nav to="/" className="navigation_items">
-        <Link to={'/'} >Home </Link>
-        <Link to={'/About'} >About </Link>
-        <Link to={'/Search'} >Search </Link>
-        <Link to={'/Login'} >Login </Link>
+      <Link to="/">Home </Link>
+      <Link to="/About">About </Link>
+      <Link to="/Search">Search </Link>
+      <Link to="/Login">Login </Link>
     </nav>
     <div className="navigation-inputbox">
       <input
