+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './components/App';
-import * as serviceWorker from './serviceWorker';
+import * as serviceWorker from './components/serviceWorker';
 
 ReactDOM.render(<App />, document.getElementById('root'));
